backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
idna==3.10
isodate==0.7.2
lxml==5.3.0
packaging==24.1
pillow==11.0.0
pipenv==2024.3.1
platformdirs==4.3.6
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==75.3.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
zeep==4.3.1
python-dotenv==1.0.1
django-jalali==6.2.1
django-jalali-date==1.1.2
persiantools==4.0.2